Refs #21321: Split in two files and move eprosima::fastdds::Time_t to eprosima::fastdds::dds::Time_t
