+++ include/fastdds/dds/core/Time_t.hpp
@@ -1,4 +1,4 @@
-// Copyright 2019 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+// Copyright 2024 Proyectos y Sistemas de Mantenimiento SL (eProsima).
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,10 +16,9 @@
  * @file Time_t.hpp
  */
 
-#ifndef FASTDDS_RTPS_COMMON__TIME_T_HPP
-#define FASTDDS_RTPS_COMMON__TIME_T_HPP
+#ifndef FASTDDS_DDS_CORE__TIME_T_HPP
+#define FASTDDS_DDS_CORE__TIME_T_HPP
 
-#include <fastdds/dds/core/Time_t.hpp>
 #include <fastdds/fastdds_dll.hpp>
 
 #include <cmath>
@@ -28,26 +27,29 @@
 
 namespace eprosima {
 namespace fastdds {
-namespace rtps {
+namespace dds {
 
 /**
- * Structure Time_t, used to describe times at RTPS protocol.
- * @ingroup COMMON_MODULE
+ * Structure Time_t, used to describe times at a DDS level.
  */
-class FASTDDS_EXPORTED_API Time_t
+struct FASTDDS_EXPORTED_API Time_t
 {
-public:
+    static constexpr int32_t INFINITE_SECONDS = 0x7fffffff;
+    static constexpr uint32_t INFINITE_NANOSECONDS = 0xffffffffu;
+
+    int32_t seconds;
+    uint32_t nanosec;
 
     //! Default constructor. Sets values to zero.
-    Time_t() = default;
+    Time_t();
 
     /**
      * @param sec Seconds
-     * @param frac Fraction of second
+     * @param nsec Nanoseconds
      */
     Time_t(
             int32_t sec,
-            uint32_t frac);
+            uint32_t nsec);
 
     /**
      * @param sec Seconds. The fractional part is converted to nanoseconds.
@@ -55,70 +57,20 @@ public:
     Time_t(
             long double sec);
 
-    /**
-     * @param time fastdds::dds::Time_t, aka. dds::Duration_t.
-     */
-    Time_t(
-            const eprosima::fastdds::dds::Time_t& time);
+    void fraction(
+            uint32_t frac);
+
+    uint32_t fraction() const;
 
     /**
      *  Returns stored time as nanoseconds (including seconds)
      */
     int64_t to_ns() const;
 
-    /**
-     *  @param nanosecs Stores given time as nanoseconds (including seconds)
-     */
-    void from_ns(
-            int64_t nanosecs);
-
-    /**
-     * Retrieve the seconds field.
-     */
-    int32_t seconds() const;
-
-    /**
-     * Retrieve the seconds field by ref.
-     */
-    int32_t& seconds();
-
-    /**
-     * Sets seconds field.
-     */
-    void seconds(
-            int32_t sec);
-
-    /**
-     * Retrieve the nanosec field.
-     */
-    uint32_t nanosec() const;
-
-    /**
-     * Sets nanoseconds field and updates the fraction.
-     */
-    void nanosec(
-            uint32_t nanos);
-
-    /**
-     * Retrieve the fraction field.
-     */
-    uint32_t fraction() const;
-
-    /**
-     * Retrieve the fraction field by ref.
-     */
-    uint32_t& fraction();
-
-    /**
-     * Sets fraction field and updates the nanoseconds.
-     */
-    void fraction(
-            uint32_t frac);
-
-    eprosima::fastdds::dds::Duration_t to_duration_t() const;
-
-    void from_duration_t(
-            const eprosima::fastdds::dds::Duration_t& duration);
+    inline bool is_infinite() const noexcept
+    {
+        return is_infinite(*this);
+    }
 
     /**
      * Fills a Time_t struct with a representation of the current time.
@@ -128,23 +80,15 @@ public:
     static void now(
             Time_t& ret);
 
-private:
+    static inline constexpr bool is_infinite(
+            const Time_t& t) noexcept
+    {
+        return (INFINITE_SECONDS == t.seconds) || (INFINITE_NANOSECONDS == t.nanosec);
+    }
 
-    //!Seconds
-    int32_t seconds_ = 0;
-
-    //!Fraction of second (1 fraction = 1/(2^32) seconds)
-    uint32_t fraction_ = 0;
-
-    //!Nanoseconds
-    uint32_t nanosec_ = 0;
-
-    void set_fraction(
-            uint32_t frac);
-
-    void set_nanosec(
-            uint32_t nanos);
 };
+
+using Duration_t = Time_t;
 
 #ifndef DOXYGEN_SHOULD_SKIP_THIS_PUBLIC
 
@@ -158,11 +102,11 @@ static inline bool operator ==(
         const Time_t& t1,
         const Time_t& t2)
 {
-    if (t1.seconds() != t2.seconds())
+    if (t1.seconds != t2.seconds)
     {
         return false;
     }
-    if (t1.fraction() != t2.fraction())
+    if (t1.nanosec != t2.nanosec)
     {
         return false;
     }
@@ -179,11 +123,11 @@ static inline bool operator !=(
         const Time_t& t1,
         const Time_t& t2)
 {
-    if (t1.seconds() != t2.seconds())
+    if (t1.seconds != t2.seconds)
     {
         return true;
     }
-    if (t1.fraction() != t2.fraction())
+    if (t1.nanosec != t2.nanosec)
     {
         return true;
     }
@@ -200,17 +144,17 @@ static inline bool operator <(
         const Time_t& t1,
         const Time_t& t2)
 {
-    if (t1.seconds() < t2.seconds())
+    if (t1.seconds < t2.seconds)
     {
         return true;
     }
-    else if (t1.seconds() > t2.seconds())
+    else if (t1.seconds > t2.seconds)
     {
         return false;
     }
     else
     {
-        if (t1.fraction() < t2.fraction())
+        if (t1.nanosec < t2.nanosec)
         {
             return true;
         }
@@ -231,17 +175,17 @@ static inline bool operator >(
         const Time_t& t1,
         const Time_t& t2)
 {
-    if (t1.seconds() > t2.seconds())
+    if (t1.seconds > t2.seconds)
     {
         return true;
     }
-    else if (t1.seconds() < t2.seconds())
+    else if (t1.seconds < t2.seconds)
     {
         return false;
     }
     else
     {
-        if (t1.fraction() > t2.fraction())
+        if (t1.nanosec > t2.nanosec)
         {
             return true;
         }
@@ -262,17 +206,17 @@ static inline bool operator <=(
         const Time_t& t1,
         const Time_t& t2)
 {
-    if (t1.seconds() < t2.seconds())
+    if (t1.seconds < t2.seconds)
     {
         return true;
     }
-    else if (t1.seconds() > t2.seconds())
+    else if (t1.seconds > t2.seconds)
     {
         return false;
     }
     else
     {
-        if (t1.fraction() <= t2.fraction())
+        if (t1.nanosec <= t2.nanosec)
         {
             return true;
         }
@@ -293,17 +237,17 @@ static inline bool operator >=(
         const Time_t& t1,
         const Time_t& t2)
 {
-    if (t1.seconds() > t2.seconds())
+    if (t1.seconds > t2.seconds)
     {
         return true;
     }
-    else if (t1.seconds() < t2.seconds())
+    else if (t1.seconds < t2.seconds)
     {
         return false;
     }
     else
     {
-        if (t1.fraction() >= t2.fraction())
+        if (t1.nanosec >= t2.nanosec)
         {
             return true;
         }
@@ -318,46 +262,8 @@ inline std::ostream& operator <<(
         std::ostream& output,
         const Time_t& t)
 {
-    return output << t.seconds() << "." << t.nanosec();
-}
-
-inline std::istream& operator >>(
-        std::istream& input,
-        Time_t& t)
-{
-    std::istream::sentry s(input);
-
-    if (s)
-    {
-        char point;
-        int32_t sec = 0;
-        uint32_t nano = 0;
-        std::ios_base::iostate excp_mask = input.exceptions();
-
-        try
-        {
-            input.exceptions(excp_mask | std::ios_base::failbit | std::ios_base::badbit);
-
-            input >> sec;
-            input >> point >> nano;
-            // nano could not be bigger or equal than 1 sec
-            if ( point != '.' || nano >= 1000000000 )
-            {
-                input.setstate(std::ios_base::failbit);
-                nano = 0;
-            }
-        }
-        catch (std::ios_base::failure& )
-        {
-        }
-
-        t.seconds(sec);
-        t.nanosec(nano);
-
-        input.exceptions(excp_mask);
-    }
-
-    return input;
+    long double t_aux = t.seconds + (((long double)t.nanosec) / 1000000000ULL);
+    return output << t_aux;
 }
 
 /**
@@ -370,10 +276,10 @@ static inline Time_t operator +(
         const Time_t& ta,
         const Time_t& tb)
 {
-    Time_t result(ta.seconds() + tb.seconds(), ta.fraction() + tb.fraction());
-    if (result.fraction() < ta.fraction()) // Overflow is detected by any of them
+    Time_t result(ta.seconds + tb.seconds, ta.nanosec + tb.nanosec);
+    if (result.nanosec < ta.nanosec) // Overflow is detected by any of them
     {
-        ++result.seconds();
+        ++result.seconds;
     }
     return result;
 }
@@ -388,26 +294,25 @@ static inline Time_t operator -(
         const Time_t& ta,
         const Time_t& tb)
 {
-    Time_t result(ta.seconds() - tb.seconds(), ta.fraction() - tb.fraction());
-    if (result.fraction() > ta.fraction()) // Overflow is detected by ta
+    Time_t result(ta.seconds - tb.seconds, ta.nanosec - tb.nanosec);
+    if (result.nanosec > ta.nanosec) // Overflow is detected by ta
     {
-        --result.seconds();
+        --result.seconds;
     }
     return result;
 }
 
 #endif // ifndef DOXYGEN_SHOULD_SKIP_THIS_PUBLIC
 
-const Time_t c_RTPSTimeInfinite{0x7fffffff, 0xffffffff};
-const Time_t c_RTPSTimeZero{0, 0};
-const Time_t c_RTPSTimeInvalid{-1, 0xffffffff};
+//! Time_t (dds::Duration_t) representing an infinite time. DONT USE IT IN CONSTRUCTORS
+const Time_t c_TimeInfinite{Time_t::INFINITE_SECONDS, Time_t::INFINITE_NANOSECONDS};
+//! Time_t (dds::Duration_t) representing a zero time. DONT USE IT IN CONSTRUCTORS
+const Time_t c_TimeZero{0, 0};
+//! Time_t (dds::Duration_t) representing an invalid time. DONT USE IT IN CONSTRUCTORS
+const Time_t c_TimeInvalid{-1, Time_t::INFINITE_NANOSECONDS};
 
-} // namespace rtps
+} // namespace dds
 } // namespace fastdds
 } // namespace eprosima
 
-// defines to avoid the "static initialization order fiasco"
-#define TIME_T_INFINITE_SECONDS (eprosima::fastdds::dds::Time_t::INFINITE_SECONDS)
-#define TIME_T_INFINITE_NANOSECONDS (eprosima::fastdds::dds::Time_t::INFINITE_NANOSECONDS)
-
-#endif // FASTDDS_RTPS_COMMON__TIME_T_HPP
+#endif // FASTDDS_DDS_CORE__TIME_T_HPP
